ecom/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==9.5.0
django~=4.0.4
celery~=5.3.4
xhtml2pdf~=0.2.11
requests~=2.31.0